+++ R/AllClassUnions.R
@@ -1,3 +1,5 @@
 #' @importClassesFrom S4Vectors DataFrame
-setClassUnion("data.frameOrSimilar", c("data.frame", "matrix", "DataFrame"))
+#' @importClassesFrom QFeatures QFeatures
+setClassUnion("data.frameOrSimilar", c("data.frame", "matrix", "DataFrame",
+                                       "SummarizedExperiment", "QFeatures"))
 setClassUnion("adductClass", c("character", "data.frame"))