+++ tests/testthat/test_Matched.R
@@ -2,12 +2,19 @@ q1 <- data.frame(col1 = 1:5, col2 = 6:10)
 t1 <- data.frame(col1 = 11:16, col2 = 17:22)
 q2 <- list("A", c(1,2), 1L, list(), data.frame())
 t2 <- list("A", c(1,2), 1L, list(), data.frame(), matrix(1:3))
+library(SummarizedExperiment)
+q3 <- SummarizedExperiment(assays = data.frame(matrix(NA, 5, 2)),
+                           rowData = q1,
+                           colData = data.frame(cD1 = c(NA, NA),
+                                                cD2 = c(NA, NA)))
+library(QFeatures)
+q4 <- QFeatures(list(a1 = q3, a2 = q3[4:5]))
+t3 <- t4 <- t1
 
 m <- data.frame(query_idx = c(1L, 2L, 2L, 2L, 5L),
                 target_idx = c(2L, 2L, 3L, 4L, 5L),
                 score = seq(0.5, 0.9, by = 0.1))
 
-#### maybe to remove in test_MatchedSpectra.R?
 test_that("validator work", {
     ## .validate_matches_format
     res <- .validate_matches_format(4)
@@ -80,6 +87,19 @@ test_that("Matched works", {
     expect_equal(whichQuery(mo), c(1L, 2L, 5L))
     expect_equal(whichTarget(mo), c(2L, 3L, 4L, 5L))
 
+    ## With data. query SummarizedExperiment and target data.frame
+    mo <- Matched(query = q3, target = t3,
+                  matches = data.frame(query_idx = c(1L, 2L, 2L, 2L, 5L),
+                                       target_idx = c(2L, 2L, 3L, 4L, 5L),
+                                       score = seq(0.5, 0.9, by = 0.1)))
+    expect_true(validObject(mo))
+    expect_true(length(mo) == 5)
+    expect_output(show(mo), "4 matched")
+    expect_true(is(query(mo), "SummarizedExperiment"))
+    expect_true(is(target(mo), "data.frame"))
+    expect_equal(whichQuery(mo), c(1L, 2L, 5L))
+    expect_equal(whichTarget(mo), c(2L, 3L, 4L, 5L))
+
 })
 
 test_that(".subset_matches_nodim and [ works", {
@@ -139,7 +159,7 @@ test_that(".subset_matches_nodim and [ works", {
     expect_equal(target(res), t1[c(2, 3, 4, 5), ])
     expect_equal(res@matches$target_idx, c(1L, 1L, 2L, 3L, 4L))
 
-    ## query and target: lists
+    #### query and target: lists
     mo <- Matched(
         q2, t2, matches = data.frame(query_idx = c(1L, 2L, 2L, 2L, 5L),
                                      target_idx = c(2L, 2L, 3L, 4L, 5L),
@@ -184,9 +204,56 @@ test_that(".subset_matches_nodim and [ works", {
     res <- pruneTarget(res)
     expect_equal(target(res), t2[c(2, 3, 4, 5)])
     expect_equal(res@matches$target_idx, c(1L, 1L, 2L, 3L, 4L))
+
+    #### query SummarizedExperiment and target data.frame
+    mo <- Matched(
+        q3, t3, matches = data.frame(query_idx = c(1L, 2L, 2L, 2L, 5L),
+                                     target_idx = c(2L, 2L, 3L, 4L, 5L),
+                                     score = seq(0.5, 0.9, by = 0.1)))
+
+    res <- .subset_matches_nodim(mo, 2)
+    expect_true(length(res) == 1)
+    expect_equal(res@matches$query_idx, c(1L, 1L, 1L))
+    expect_equal(res@matches$target_idx, mo@matches$target_idx[c(2, 3, 4)])
+    expect_equal(res@matches$score, mo@matches$score[c(2, 3, 4)])
+    expect_equal(res@query, mo@query[2, ])
+    expect_equal(res@target, mo@target)
+
+
+    expect_error(.subset_matches_nodim(mo, 12), "out-of-bounds")
+
+    res <- .subset_matches_nodim(mo, c(1, 5))
+    expect_equal(res@query, mo@query[c(1, 5), ])
+    expect_true(length(res) == 2)
+    expect_equal(res@matches$score, mo@matches$score[c(1, 5)])
+
+    ## arbitrary order
+    res <- .subset_matches_nodim(mo, c(3, 2, 1, 1, 4))
+    expect_equal(query(res), query(mo)[c(3, 2, 1, 1, 4), ])
+    expect_equal(target(res), target(mo))
+    expect_equal(res@matches$query_idx, c(2L, 2L, 2L, 3L, 4L))
+    expect_equal(res@matches$score, c(0.6, 0.7, 0.8, 0.5, 0.5))
+
+    res <- mo[]
+    expect_equal(res, mo)
+    res <- mo[c(TRUE, FALSE)]
+    expect_equal(res@matches$query_idx, 1L)
+    expect_equal(res@matches$target_idx, 2L)
+    expect_equal(res@matches$score, 0.5)
+
+    ## All works even after subsetting and pruning.
+    mo <- Matched(
+        q3, t3, matches = data.frame(query_idx = c(1L, 2L, 2L, 2L, 5L),
+                                     target_idx = c(2L, 2L, 3L, 4L, 5L),
+                                     score = seq(0.5, 0.9, by = 0.1)))
+
+    res <- mo[whichQuery(mo)]
+    expect_equal(query(res), q3[whichQuery(mo), ])
+    res <- pruneTarget(res)
+    expect_equal(target(res), t3[c(2, 3, 4, 5), ])
+    expect_equal(res@matches$target_idx, c(1L, 1L, 2L, 3L, 4L))
 })
 
-#### maybe to remove in test_MatchedSpectra.R?
 test_that(".fill_index works", {
     res <- .fill_index(1:5, integer())
     expect_equal(res, 1:5)
@@ -236,6 +303,14 @@ test_that("$,Matched and .dollar works", {
     tmp <- t2[c(2,2,3,4, NA, NA, 5)]
     tmp[5:6] <- NA
     expect_equal(mo$target, tmp)
+
+    mo <- Matched(
+        q3, t3, matches = data.frame(query_idx = c(1L, 2L, 2L, 2L, 5L),
+                                     target_idx = c(2L, 2L, 3L, 4L, 5L),
+                                     score = seq(0.5, 0.9, by = 0.1)))
+    expect_equal(mo$score, c(0.5, 0.6, 0.7, 0.8, NA, NA, 0.9))
+    expect_equal(mo$col1, rowData(q3)[c(1, 2, 2, 2, 3, 4, 5), "col1"])
+    expect_equal(mo$target_col1, t3[c(2,2,3,4, NA, NA, 5), "col1"])
 })
 
 test_that(".dollar works with unordered @matches", {
@@ -293,6 +368,12 @@ test_that("colnames,Matched works", {
     expect_equal(colnames(mo), c("query",
                                  paste0("target_", colnames(t1)), "score"))
 
+    mo <- Matched(
+        q3, t3, matches = data.frame(query_idx = c(1L, 2L, 2L, 2L, 5L),
+                                     target_idx = c(2L, 2L, 3L, 4L, 5L),
+                                     score = seq(0.5, 0.9, by = 0.1)))
+    expect_equal(colnames(mo), c(colnames(rowData(q3)),
+                                 paste0("target_", colnames(t3)), "score"))
 })
 
 test_that("matchedData,Matched works", {
@@ -404,6 +485,57 @@ test_that("matchedData,Matched works", {
     res <- matchedData(mo, columns = c("score"))
     expect_equal(colnames(res), c("score"))
     expect_equal(res$score, mo$score)
+
+    #### query SummarizedExoeriment and target data.frame
+    mo <- Matched(
+        q3, t3, matches = data.frame(query_idx = c(1L, 2L, 2L, 2L, 5L),
+                                     target_idx = c(2L, 2L, 3L, 4L, 5L),
+                                     score = seq(0.5, 0.9, by = 0.1)))
+    res <- matchedData(mo)
+    expect_equal(res$col2, mo$col2)
+    expect_equal(res$target_col2, mo$target_col2)
+    expect_equal(res$score, mo$score)
+
+    expect_error(matchedData(mo, columns = "other"), "other not available")
+
+    ## Only query object variables
+    res <- matchedData(mo, columns = c("col1", "col2"))
+    expect_equal(colnames(res), c("col1", "col2"))
+    expect_equal(res$col1, mo$col1)
+
+    ## Only target object variables
+    res <- matchedData(mo, columns = c("target_col1", "target_col2"))
+    expect_equal(colnames(res), c("target_col1", "target_col2"))
+    expect_equal(res$target_col1, mo$target_col1)
+
+    ## Only matches
+    res <- matchedData(mo, columns = c("score"))
+    expect_equal(colnames(res), c("score"))
+    expect_equal(res$score, mo$score)
+
+    ## A MatchedSummarizedExperiment with no matching target elements
+    mo <- Matched(q3, t3, matches = data.frame(query_idx = integer(),
+                                               target_idx = integer(),
+                                               score = numeric()))
+    res <- matchedData(mo)
+    expect_equal(res$col2, mo$col2)
+    expect_equal(res$target_col2, mo$target_col2)
+    expect_equal(res$score, mo$score)
+
+    ## Only query object variables
+    res <- matchedData(mo, columns = c("col1", "col2"))
+    expect_equal(colnames(res), c("col1", "col2"))
+    expect_equal(res$col1, mo$col1)
+
+    ## Only target object variables
+    res <- matchedData(mo, columns = c("target_col1", "target_col2"))
+    expect_equal(colnames(res), c("target_col1", "target_col2"))
+    expect_equal(res$target_col1, mo$target_col1)
+
+    ## Only matches
+    res <- matchedData(mo, columns = c("score"))
+    expect_equal(colnames(res), c("score"))
+    expect_equal(res$score, mo$score)
 })
 
 test_that("pruneTarget,Matched works", {
@@ -443,6 +575,22 @@ test_that("pruneTarget,Matched works", {
     res <- pruneTarget(mo)
     expect_equal(matchedData(res), matchedData(mo))
     expect_true(length(res@target) < length(mo@target))
+
+    #### query SummarizedExperiment and target data.frame
+    mo <- Matched(
+        q3, t3, matches = data.frame(query_idx = c(1L, 2L, 2L, 2L, 5L),
+                                     target_idx = c(2L, 2L, 3L, 4L, 5L),
+                                     score = seq(0.5, 0.9, by = 0.1)))
+    res <- pruneTarget(mo)
+    expect_equal(matchedData(res), matchedData(mo))
+    expect_true(nrow(res@target) < nrow(mo@target))
+
+    mo <- Matched(q3, t3, matches = data.frame(query_idx = integer(),
+                                               target_idx = integer(),
+                                               score = numeric()))
+    res <- pruneTarget(mo)
+    expect_equal(matchedData(res), matchedData(mo))
+    expect_true(nrow(res@target) < nrow(mo@target))
 })
 
 test_that("filterMatches,Matched works", {
@@ -490,17 +638,52 @@ test_that("filterMatches,Matched works", {
 
     #### query and target vectors (multiple matches corresponding to a given
     #### couple of query and target input values)
-    q3 <- c(1, 1, 2, 3, 4)
-    t3 <- c("A", "A", "A", "C", "D")
     mo <- Matched(
-        q3, t3, matches = data.frame(query_idx = c(1L, 2L, 4L, 5L),
-                                     target_idx = c(1L, 2L, 4L, 5L),
-                                     score = seq(0.5, 0.8, by = 0.1)))
+        c(1, 1, 2, 3, 4), c("A", "A", "A", "C", "D"),
+        matches = data.frame(query_idx = c(1L, 2L, 4L, 5L),
+                             target_idx = c(1L, 2L, 4L, 5L),
+                             score = seq(0.5, 0.8, by = 0.1)))
     queryValue <- c(1, 3)
     targetValue <- c("A", "C")
     mosub <- filterMatches(mo, queryValue = queryValue,
                          targetValue = targetValue)
     expect_equal(mosub@matches, mo@matches[c(1, 2, 3), ])
+
+    #### query SummarizedExperiment and target data.frame
+    mo <- Matched(
+        q3, t3, matches = data.frame(query_idx = c(1L, 2L, 2L, 2L, 5L),
+                                     target_idx = c(2L, 2L, 3L, 4L, 5L),
+                                     score = seq(0.5, 0.9, by = 0.1)))
+    ## out of bounds indexes
+    expect_error(filterMatches(mo, index = c(1, 10)), "out-of-bounds")
+    ## no index : every match is removed
+    idxs <- integer(0)
+    mosub <- filterMatches(mo, index = idxs)
+    expect_equal(mosub@matches, mo@matches[idxs, ])
+    expect_equal(query(mosub), query(mo))
+    expect_equal(target(mosub), target(mo))
+    ## in range indexes
+    idxs <- c(1, 3, 5)
+    mosub <- filterMatches(mo, index = idxs)
+    expect_equal(mosub@matches, mo@matches[idxs, ])
+    expect_equal(query(mosub), query(mo))
+    expect_equal(target(mosub), target(mo))
+    ## keep matches based on query and target input values
+    queryValue <- c(rowData(q3)[mo@matches[idxs, "query_idx"], "col1"], -1, - 2)
+    targetValue <- c(t3[mo@matches[idxs, "target_idx"], "col2"], -2, -1)
+    mosub <- filterMatches(mo, queryValue = queryValue,
+                           targetValue = targetValue,
+                           queryColname = "col1", targetColname = "col2")
+    expect_equal(mosub@matches, mo@matches[idxs, ])
+    expect_equal(query(mosub), query(mo))
+    expect_equal(target(mosub), target(mo))
+    ## no matches corresponding to the input values
+    mosub <- filterMatches(mo, queryValue = queryValue + 100,
+                           targetValue = targetValue + 100,
+                           queryColname = "col1", targetColname = "col2")
+    expect_equal(mosub@matches, data.frame(query_idx = integer(),
+                                           target_idx = integer(),
+                                           score = numeric()))
 })
 
 test_that("addMatches,Matched works", {
@@ -530,10 +713,10 @@ test_that("addMatches,Matched works", {
     expect_equal(target(moadd), target(mo))
 
     #### query vector and target data.frame
-    mo <- Matched(
-        q1[, "col1"], t1, matches = data.frame(query_idx = c(1L, 2L, 2L, 2L, 5L),
-                                     target_idx = c(2L, 2L, 3L, 4L, 5L),
-                                     score = seq(0.5, 0.9, by = 0.1)))
+    mo <- Matched(q1[, "col1"], t1,
+                  matches = data.frame(query_idx = c(1L, 2L, 2L, 2L, 5L),
+                                       target_idx = c(2L, 2L, 3L, 4L, 5L),
+                                       score = seq(0.5, 0.9, by = 0.1)))
     moadd <- addMatches(mo, queryValue = c(q1[c(3, 5), "col1"], 50),
                         targetValue = c(t1[c(1, 6), "col2"], 50),
                         targetColname = "col2",
@@ -552,6 +735,19 @@ test_that("addMatches,Matched works", {
     expect_equal(moadd@matches$score, c(mo@matches$score, 1, 1.1))
     expect_equal(query(moadd), query(mo))
     expect_equal(target(moadd), target(mo))
+
+    #### query SummarizedExperiment and target data.frame
+    mo <- Matched(
+        query = q3, target = t3,
+        matches = data.frame(query_idx = c(1L, 2L, 2L, 2L, 5L),
+                             target_idx = c(2L, 2L, 3L, 4L, 5L),
+                             score = seq(0.5, 0.9, by = 0.1)))
+
+    moadd <- addMatches(mo, queryValue = 2L, targetValue = 1L, isIndex = TRUE,
+                        score = data.frame(score = 100))
+    expect_true(matches(moadd)$query_idx[6] == 2)
+    expect_true(matches(moadd)$target_idx[6] == 1)
+    expect_true(matches(moadd)$score[6] == 100)
 })
 
 test_that(".extract_elements works", {
@@ -574,10 +770,46 @@ test_that(".extract_elements works", {
 test_that(".cnt works", {
     t <- data.frame(a = 1:5, b = 2:6)
     expect_equal(.cnt(t), c("target_a", "target_b"))
-    
+
     t <- 1:5
     expect_equal(.cnt(t), "target")
 
     t <- array(data = 1, dim = c(1, 1, 1))
     expect_error(.cnt(t), "unsupported")
+})
+
+test_that(".validate_assay works", {
+    expect_identical(.validate_assay(q4, c("a1", "a2")),
+                     "`queryAssay` must be `character(1)`")
+    expect_identical(.validate_assay(q4, "a3", "target"),
+                     "No assay in `target` with name \"a3\"")
+    expect_null(.validate_assay(q4, "a1"))
+})
+
+test_that(".objectToMatch works", {
+    expect_identical(.objectToMatch(q1), q1)
+    expect_identical(.objectToMatch(q2), q2)
+    expect_identical(.objectToMatch(q3), rowData(q3))
+    expect_error(.objectToMatch(q4), "has to be provided")
+    expect_error(.objectToMatch(q4, c("a1", "a2")), "must be `character")
+    expect_error(.objectToMatch(q4, c("a3")), "with name")
+    expect_identical(.objectToMatch(q4, c("a1")), rowData(q4[["a1"]]))
+    expect_identical(.objectToMatch(q1, "assayname", "col2"), q1[, "col2"])
+    expect_identical(.objectToMatch(q3, colnames = "col2"),
+                     rowData(q3)[, "col2"])
+    expect_identical(.objectToMatch(q4, "a1", "col2"),
+                     rowData(q4[["a1"]])[, "col2"])
+    expect_error(.objectToMatch(q1, "a1", c("col1", "col3")), "Missing column")
+    expect_error(.objectToMatch(q3, "a1", c("col1", "col3")), "Missing column")
+    expect_error(.objectToMatch(q4, "a1", c("col1", "col3")), "Missing column")
+})
+
+test_that(".subset_qt works", {
+    i <- c(1, 3)
+    expect_identical(.subset_qt(q1, i = i), q1[i, ])
+    expect_identical(.subset_qt(q2, i = i), q2[i])
+    expect_identical(.subset_qt(q3, i = i), (q3)[i, ])
+    res <- .subset_qt(q4, "a1", i = i)
+    expect_is(res, "QFeatures")
+    expect_equal(res[["a1"]], q4[["a1"]][i, ])
 })